ya se migran los datos locales a la db

+++ src/pages/Login.jsx
@@ -7,7 +7,11 @@ import { useForm } from "react-hook-form";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router";
 import Swal from "sweetalert2";
-import { deleteCart, getCartThunk, migrateLocalCart } from "../store/slices/cart.slice";
+import {
+  deleteCart,
+  getCartThunk,
+  migrateLocalCart,
+} from "../store/slices/cart.slice";
 import { setUser } from "../store/slices/user.slice";
 
 const Login = () => {
@@ -15,19 +19,15 @@ const Login = () => {
   const { register, handleSubmit } = useForm();
   const dispatch = useDispatch();
   const navigate = useNavigate();
-const cart = useSelector(state => state.cart)
-
+  const cart = useSelector((state) => state.cart);
 
   const swalWithBootstrapButtons = Swal.mixin({
     customClass: {
-      confirmButton: 'btn btn-success',
-      cancelButton: 'btn btn-danger'
+      confirmButton: "btn btn-success",
+      cancelButton: "btn btn-danger",
     },
-    buttonsStyling: false
-  })
-  
-
-
+    buttonsStyling: false,
+  });
 
   const submit = (data) => {
     axios
@@ -42,65 +42,52 @@ const cart = useSelector(state => state.cart)
 
         navigate("/");
 
+        /// soolo si hay elementos en carrito
+        if (cart.length !== 0) {
+          swalWithBootstrapButtons
+            .fire({
+              title: "Keep products in cart?",
+              text: "There is a list of products in your cart before login, do you want to keep those products in your cart?",
+              icon: "warning",
+              showCancelButton: true,
+              confirmButtonText: "Yes, keep cart",
+              cancelButtonText: "No, delete cart",
+              reverseButtons: true,
+            })
+            .then((result) => {
+              if (result.isConfirmed) {
+                const elements = cart.filter(
+                  (item, index) => cart.indexOf(item) === index
+                );
 
+                const productsCart = elements.map((product) => {
+                  let count = 0;
+                  cart.map((item) => {
+                    if (product.id === item.id) {
+                      count += 1;
+                    }
+                  });
+                  return { product, count };
+                });
 
-/// soolo si hay elementos en carrito
-if(cart.length !== 0) {
-  swalWithBootstrapButtons.fire({
-    title: 'Keep products in cart?',
-    text: "There is a list of products in your cart before login, do you want to keep those products in your cart?",
-    icon: 'warning',
-    showCancelButton: true,
-    confirmButtonText: 'Yes, keep cart',
-    cancelButtonText: 'No, delete cart',
-    reverseButtons: true
-  }).then((result) => {
-    if (result.isConfirmed) {
+                dispatch(migrateLocalCart(productsCart));
 
-      const elements = cart.filter((item, index) => cart.indexOf(item) === index);
+                swalWithBootstrapButtons.fire(
+                  "Deleted!",
+                  "Your file has been deleted.",
+                  "success"
+                );
+              } else if (result.dismiss === Swal.DismissReason.cancel) {
+                dispatch(deleteCart());
 
-  const productsCart = elements.map((product) => {
-    let count = 0;
-    cart.map((item) => {
-      if (product.id === item.id) {
-        count += 1;
-      }
-    });
-    return { product, count };
-  });
-  
-
-dispatch(migrateLocalCart(productsCart))
-
-
-
-
-      swalWithBootstrapButtons.fire(
-        'Deleted!',
-        'Your file has been deleted.',
-        'success'
-      )
-    } else if (result.dismiss === Swal.DismissReason.cancel) {
-
-
-      dispatch(deleteCart())
-
-      swalWithBootstrapButtons.fire(
-        'Cancelled',
-        'Your imaginary file is safe :)',
-        'error'
-      )
-    }
-  })
-}
-
-        
-
-
-
-
-
-
+                swalWithBootstrapButtons.fire(
+                  "Cancelled",
+                  "Your imaginary file is safe :)",
+                  "error"
+                );
+              }
+            });
+        }
       })
       .catch((err) => {
         Swal.fire({
